refactor: winningInfo에서 count, hasBonus check
